test-sso/pyproject.toml
[project]
name = "test-azure-entra-sso"
version = "0.1.0"
description = "Azure Entra ID SSO testing for Neo4j Aura"
requires-python = ">=3.10"
dependencies = [
    "msal>=1.28.0",
    "neo4j>=5.5.0",
    "python-dotenv>=1.2.1",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]
